Completed table page.
