+++ src/App.css
@@ -68,3 +68,35 @@ button {
   color: aliceblue;
   cursor: pointer;
 }
+
+table {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+thead {
+  width: 100%;
+}
+
+tr {
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+}
+
+td {
+  font-size: 0.8rem;
+  text-align: center;
+}
+
+th, td {
+  width: 10%;
+}
+
+table, th, td {
+  border:1px solid #fff;
+  color: #fff;
+}
