+++ src/redux/configureStore.js
@@ -2,9 +2,11 @@ import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
 import leaguesReducer from './League/league';
+import tableReducer from './Table/table';
 
 const reducer = combineReducers({
   leaguesReducer,
+  tableReducer,
   // additional reducers could be added here
 });
